Move disk encryption under generic storage configuration step
The motivation behind this was to prepare the ground to introduce more
configuration options, like `swap` option.
