+++ test/check-basic
@@ -65,11 +65,8 @@ class TestBasic(VirtInstallMachineCase):
         # Test going back
         steps = [
             i.steps.ACCOUNTS,
-            i.steps.DISK_CONFIGURATION,
-            i.steps.DISK_ENCRYPTION,
-            i.steps.DISK_CONFIGURATION,
+            i.steps.STORAGE_CONFIGURATION,
             i.steps.INSTALLATION_METHOD,
-            i.steps.WELCOME,
             i.steps.WELCOME,
         ]
 
@@ -83,7 +80,7 @@ class TestBasic(VirtInstallMachineCase):
         i.reach_on_sidebar(i.steps.WELCOME)
         for step in [
             i.steps.ACCOUNTS,
-            i.steps.DISK_CONFIGURATION,
+            i.steps.STORAGE_CONFIGURATION,
             i.steps.INSTALLATION_METHOD,
             i.steps.REVIEW,
         ]:
