+++ test/check-review
@@ -63,7 +63,7 @@ class TestReview(VirtInstallMachineCase):
             ignore=pixel_tests_ignore,
         )
 
-        i.reach_on_sidebar(i.steps.DISK_ENCRYPTION)
+        i.reach_on_sidebar(i.steps.STORAGE_CONFIGURATION)
         s.set_encryption_selected(True)
         strong_password = "Rwce82ybF7dXtCzFumanchu!!!!!!!!"
         p.set_password(strong_password)
