+++ test/check-storage-encryption
@@ -128,7 +128,7 @@ class TestStorageEncryption(VirtInstallMachineCase, StorageHelpers):
         p = Password(b, s.encryption_id_prefix)
 
         i.open()
-        i.reach(i.steps.DISK_ENCRYPTION)
+        i.reach(i.steps.STORAGE_CONFIGURATION)
         # Disk Encryption / password screen
 
         encrypt = True
